+++ csm/conf/setup.py
@@ -164,36 +164,6 @@ class Setup:
         Setup._run_cmd("chown -R "+self._user+":"+self._user+" "+os.path.join(home_dir, const.SSH_DIR))
         Setup._run_cmd("chmod 400 "+os.path.join(const.CSM_USER_HOME, const.SSH_PRIVATE_KEY))
 
-    def _config_user(self, reset=False):
-        """
-        Check user already exist and create if not exist
-        If reset true then delete user
-        """
-        Log.info("Check user already exist and create if not exist. reset flag: {reset}")
-        if not reset:
-            if not self._is_user_exist():
-                Setup._run_cmd("useradd -d "+const.CSM_USER_HOME+" -p "+self._password+" "+ self._user)
-                Setup._run_cmd("usermod -aG wheel " + self._user)
-                if not self._is_user_exist():
-                    raise CsmSetupError("Unable to create %s user" % self._user)
-                node_name = SaltWrappers.get_salt_call(const.GRAINS_GET, 'id', 'log')
-                primary = SaltWrappers.get_salt_call(const.GRAINS_GET, 'roles', 'log')
-                if ( node_name is None or const.PRIMARY_ROLE in primary):
-                    self._passwordless_ssh(const.CSM_USER_HOME)
-                nodes = SaltWrappers.get_salt_call(const.PILLAR_GET, const.NODE_LIST_KEY, 'log')
-                if ( primary and const.PRIMARY_ROLE in primary and nodes is not None and len(nodes) > 1 ):
-                    nodes.remove(node_name)
-                    for node in nodes:
-                        if (self._check_if_dir_exist_remote_host(const.CSM_USER_HOME, node)):
-                            Setup._run_cmd("scp -pr "+os.path.join(const.CSM_USER_HOME, const.SSH_DIR)+" "+
-                                      node+":"+const.CSM_USER_HOME)
-                            Setup._run_cmd(" ssh "+node+" chown -R "+self._user+":"+self._user+" "+
-                                                 os.path.join(const.CSM_USER_HOME, const.SSH_DIR) )
-        else:
-            if self._is_user_exist():
-                Setup._run_cmd("userdel -r " +self._user)
-        if self._is_user_exist() and Setup._is_group_exist(const.HA_CLIENT_GROUP):
-            Setup._run_cmd(f"usermod -a -G {const.HA_CLIENT_GROUP}  {self._user}")
 
     def _config_user_permission_set(self, bundle_path, crt, key):
         """
@@ -634,87 +604,8 @@ class Setup:
             Log.error(f"Refresh Context: Resolving of alerts failed. {ex}")
             raise CsmSetupError(f"Refresh Context: Resolving of alerts failed. {ex}")
 
-    def set_unsupported_feature_info(self):
-        """
-        This method stores CSM unsupported features in two ways:
-        1. It first gets all the unsupported features lists of the components,
-        which CSM interacts with. Add all these features as CSM unsupported
-        features. The list of components, CSM interacts with, is
-        stored in csm.conf file. So if there is change in name of any
-        component, csm.conf file must be updated accordingly.
-        2. Installation/envioronment type and its mapping with CSM unsupported
-        features are maintained in unsupported_feature_schema. Based on the
-        installation/environment type received as argument, CSM unsupported
-        features can be stored.
-        """
-        def get_component_list_from_features_endpoints():
-            feature_endpoints = Json(const.FEATURE_ENDPOINT_MAPPING_SCHEMA).load()
-            component_list = [feature for v in feature_endpoints.values() for feature in v.get(const.DEPENDENT_ON)]
-            return list(set(component_list))
 
-        try:
-            Log.info("Set unsupported feature list to ES")
-            self._setup_info  = self.get_data_from_provisioner_cli(const.GET_SETUP_INFO)
-            unsupported_feature_instance = unsupported_features.UnsupportedFeaturesDB()
-            self._loop = asyncio.get_event_loop()
-            components_list = get_component_list_from_features_endpoints()
-            unsupported_features_list = []
-            for component in components_list:
-                unsupported = self._loop.run_until_complete(
-                    unsupported_feature_instance.get_unsupported_features(component_name=component))
-                for feature in unsupported:
-                    unsupported_features_list.append(feature.get(const.FEATURE_NAME))
 
-            csm_unsupported_feature = Json(const.UNSUPPORTED_FEATURE_SCHEMA).load()
-
-            for setup in csm_unsupported_feature[const.SETUP_TYPES]:
-                if setup[const.NAME] == self._setup_info[const.STORAGE_TYPE]:
-                    unsupported_features_list.extend(setup[const.UNSUPPORTED_FEATURES])
-            unsupported_features_list = list(set(unsupported_features_list))
-            unique_unsupported_features_list = list(filter(None, unsupported_features_list))
-            if unique_unsupported_features_list:
-                self._loop.run_until_complete(unsupported_feature_instance.store_unsupported_features(
-                    component_name=str(const.CSM_COMPONENT_NAME), features=unique_unsupported_features_list))
-            else:
-                Log.info("Unsupported features list is empty.")
-        except Exception as e_:
-            Log.error(f"Error in storing unsupported features: {e_}")
-            # TODO: Suppressing the error for now.
-            # raise CsmSetupError(f"Error in storing unsupported features: {e_}")
-
-    def _configure_system_auto_restart(self):
-        """
-        Check's System Installation Type an dUpdate the Service File
-        Accordingly.
-        :return: None
-        """
-        Log.info("Configuring System Auto restart")
-        is_auto_restart_required = list()
-        if self._setup_info:
-            for each_key in self._setup_info:
-                comparison_data = const.EDGE_INSTALL_TYPE.get(each_key, None)
-                #Check Key Exists:
-                if comparison_data is None:
-                    Log.warn(f"Edge Installation missing key {each_key}")
-                    continue
-                if isinstance(comparison_data, list):
-                    if self._setup_info[each_key] in comparison_data:
-                        is_auto_restart_required.append(False)
-                    else:
-                        is_auto_restart_required.append(True)
-                elif self._setup_info[each_key] == comparison_data:
-                    is_auto_restart_required.append(False)
-                else:
-                    is_auto_restart_required.append(True)
-        else:
-            Log.warn("Setup info does not exist.")
-            is_auto_restart_required.append(True)
-        if any(is_auto_restart_required):
-            Log.debug("Updating All setup file for Auto Restart on "
-                             "Failure")
-            Setup._update_service_file("#< RESTART_OPTION >",
-                                      "Restart=on-failure")
-            Setup._run_cmd("systemctl daemon-reload")
 
     @staticmethod
     def _update_service_file(key, value):
